Added The EventDataAccess And The CalendarDataAccess
I add the EventDataAccess and the CalendarDataAccess and all the neccessary methods that are used in them. I also updated the UserDataAccess to fit with the changes I made to the other classes. I also implemented dependency injection for all of them and added the necessary documentation for each of the classes.