Added Home Controller + HomePage And AddCalendar Views - Konstantinos
I added the home controller and I also implemented the back end of the homepage and add calendar views. Additionally I added the method delete calendar in the homecontroller, which can be accessed from the homepage view and I also implemented the logout method in the homecontroller, which can be called from the navbar. All the navbar support for the given views has also been implemented. Additionally I improved the login and register pages and fixed their logic a bit.