Updated ViewCalendarDay And Created Authentication System -- Konstantinos
I updated the view calendar day and now the alert of event deletion works correctly. Also the column buttons, which did not work correctly now work correctly. I also noticed a bug(I will fix it next time) with the events(the events do not check from the bottom and they must also check from there). I also imlemented a very basic authentication system for the application, which uses simply a static class to check if a user is authenticated or not.