Updated The DataAccess And Improved The Notifications View --Konstantinos
I changed the database to be able to support notifications and also comments. More specifically I added into the EventDataAccess class the necessary methods for the creation, deletion and retrieval of comments and notifications. I also improved the front end for the notificationsView and refactored the code for the dataaccess logic and the models and datamodels.