Fixed A DataAccess Bug And Added The DataAccess Implementation For The Alert Status Notifications --Konstantinos
I fixed the getEvent database bug in which the getEvent was using only the Id and instead of returning one record it was returning many records with false dynamic event information(at that point the only dynamic information is the alert status). In order to fix that I created a new getEvent that also specifies the correct user and so it returns the correct dynamic information. I also changed the database and dataaccess a bit to also implement the alert status notifications for the application.