Updated The DataAccess To Support AlertStatus Notifications And Fixed A DataAccess Bug --Konstantinos
I fixed the dataaccess bug that was caused by the getEvent method. This method once called did not return one record, but all the records of that existed for that event based on the ParticipationInEvent table. Once I discovered that I removed the connection that this method had with the ParticipationInEvent table and it now only returns the static information of the event(the information that exists in the event table). I also created an overload of this method that has an additional username parameter, which should now return the static and dynamic information of the event(the information that exists in the event and in the participationInEvent tables) for the the given user correctly. I also made the necessary changes in the dataaccess for the implementation of the alertstatus notifications and also updated the database by adding the AlertNotification column in the Notification table.