Prototype Finalization -- Konstantinos
With this push the prototype is complete, which means that all functions that were required, by the assignment have been functionally completed. As it stands some parts of the application are not fully responsive and some parts are also coherent with the rest of the application. The next versions of the app will mostly focus on fixing the above issues and moving forward to the final version. Additionally we will propably add extra functionality such as persistent login, email support, ability to reset user password etc. Finally we will also try to make the application as bug free as possible, but this is a fine point to call the first and most imprtant version(MVP) of the application complete.