Added Email Functionality, Improved The Navbar And Especially The Footer And Improved Responsiveness & Concistency --Konstantinos
With this update the app email functionality, which means that the contact form in the footer is capable of sending emails to us and that for a character creation a authentication email must be sent before the registration is fully complete. Additionality we added the ResetPassword view, which using an authentication email allows the user to change their password. Additionaly the footer and the navbar were improved, now they follow the color patters of the app and footer has functionality which allow the user to contact us easier. The navbar will be updated further in a future update. Finally we made sure that this app is 100% responsive for mobile, tablet and desktop and we improved the concistency of the whole app(fonts, colors, buttons and other).