Improved Concistency And Responsibility Even More And Added The Screenshot Feature For Calendars --Konstantinos
I improved the concistency by adding sevrons in all forms for mobile and also improved responsibility for the HomePage and for the ViewNotifications page. Additionally I added the automatic screenshot feature for the calendars and the database and dataaccess requirements to support it. After this update the screenshot feature should work corectly for the HomePage.