Overhaul of the Authentication And Authorization System And Notifications Improvements --Konstantinos
I overhauled the authentication and authorization system of the application, which now uses a session system. The session system saves cookies with the necessary information on the users browser and the necessary session token on the database. With this mechanism multiple users can now use the app simultaneously since at every request the user gets authenticated and authorized. Additionally I improved the notification popups of the system, which now get updated every time a user makes a new request. I could add a background service for that, but I am not certain if this will cost a lot of resources. In the future I might do that, but for this update it remains that way.