Final Update For Now Updated ViewNotifications, Fixed Bugs & Fixed A Big Bug On The HomePage.  --Konstantinos
The app as it stands is fully funtional and this will be the final update for now. The view notifications update has been updated and now is more fleshed out(there was an updated in the database and dataacces for that). Various bugs were fixed. One significant bug fix was in the editAccount action where it stopped working because of the new authentication system. After this update it should be fixed. The biggest bug was a bug that had went unnoticed. Essentionally the paging on the homepage when the user has more than 4 calendars was not really working, because if you where on tablet and you clicked number 2 the action was interpreting that as number 2 also for the desktop, which caused the app to crash. That and smaller issues have now "hopefully" been fixed. In the days left we will monitor the app for any bugs and we will try to eliminate as much of them as possible.